requirements.txt
flask
flask_restx
flask_cors
flas_sqlalchemy
flask_mail
argon2-cffi
pyjwt